# openapi.yaml
openapi: 3.0.3
info:
  title: School Reading & Assessment Platform API
  version: 1.0.0
  description: |
    Complete REST API for the school oral reading and AI assessment platform.

    ## Authentication
    All endpoints except `/auth/register` and `/auth/login` require a JWT Bearer token
    in the `Authorization: Bearer <token>` header. Access tokens expire in 15 minutes.
    Use `/auth/refresh` to obtain a new one without re-logging in.

    ## Roles
    | Role    | Access                                                        |
    |---------|---------------------------------------------------------------|
    | admin   | School structure, user management, all assignments            |
    | teacher | Their assigned sections, lesson upload, student progress view |
    | student | Their section's lessons, reading sessions, personal dashboard |

    ## Async Assessment Flow
    1. `POST /student/sessions` — creates session record
    2. Student records audio in browser via MediaRecorder API
    3. `POST /student/sessions/{id}/submit` — uploads audio (WebM/OGG)
    4. Server streams audio to Whisper STT microservice (audio never stored)
    5. Transcript sent to Ollama LLM for assessment (BullMQ job)
    6. **Socket.io** emits `assessment.complete` to student's browser
    7. `GET /student/sessions/{id}/result` — fetch the assessment result

    ## Socket.io Events
    Connect with the JWT token as a query param: `ws://server/api/v1?token=<jwt>`

...
servers:
...
security:
  - bearerAuth: []

# ─────────────────────────────────────────────────────────────────────────────
# COMPONENTS
# ─────────────────────────────────────────────────────────────────────────────
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from POST /auth/login

  # ── Reusable Parameters ───────────────────────────────────────────────────
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number (1-indexed)
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: Results per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    ClassIdParam:
      name: classId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    SectionIdParam:
      name: sectionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    UUID:
      type: string
      format: uuid
      example: "550e8400-e29b-41d4-a716-446655440000"
    Timestamp:
      type: string
      format: date-time
      example: "2026-05-30T08:30:00Z"
...
    Pagination:
      type: object
      required: [total, page, limit, total_pages]
      properties:
        total:
          type: integer
          example: 150
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 8
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: NOT_FOUND
        message:
          type: string
          example: Resource not found
        details:
          type: object
          additionalProperties: true
          nullable: true
    FieldError:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
          example: email
        message:
          type: string
          example: must be a valid email address
    ValidationError:
      type: object
      required: [code, message, errors]
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Request validation failed
        errors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'

    # ── Auth Schemas ──────────────────────────────────────────────────────
...
    Class:
      type: object
      required: [id, name, is_active, section_count, created_at]
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
          example: Grade 5
        is_active:
          type: boolean
          example: true
        section_count:
          type: integer
          example: 3
        created_at:
          $ref: '#/components/schemas/Timestamp'
    CreateClassRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          example: Grade 5
    UpdateClassRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          example: Grade 5 (Revised)

    # ── Section Schemas ───────────────────────────────────────────────────
    Section:
      type: object
      required: [id, class_id, class_name, name, is_active, student_count, teacher_count, created_at]
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        class_id:
          $ref: '#/components/schemas/UUID'
        class_name:
          type: string
          example: Grade 5
        name:
          type: string
          example: Section A
        is_active:
          type: boolean
          example: true
        student_count:
          type: integer
          example: 32
        teacher_count:
          type: integer
          example: 2
        created_at:
          $ref: '#/components/schemas/Timestamp'
    CreateSectionRequest:
      type: object
      required: [class_id, name]
      properties:
        class_id:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
          minLength: 1
          maxLength: 100
          example: Section A
    UpdateSectionRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          example: Section A (Revised)

    # ── Assignment Schemas ────────────────────────────────────────────────
...
  responses:
    Unauthorized:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: UNAUTHORIZED
            message: Invalid or expired token
    Forbidden:
      description: Authenticated but insufficient role permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: FORBIDDEN
            message: You do not have permission to access this resource
    NotFound:
      description: Requested resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: NOT_FOUND
            message: Resource not found
    Conflict:
      description: Resource already exists or business rule violated
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: CONFLICT
            message: A section named 'Section A' already exists in Grade 5
    UnprocessableEntity:
      description: Request body or query parameters failed schema validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
...
paths:
...
  /admin/classes:
    get:
      tags: [Admin — Classes]
      summary: List all classes
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: include_inactive
          in: query
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Paginated class list
          content:
            application/json:
              schema:
                type: object
                required: [data, pagination]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Class'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [Admin — Classes]
      summary: Create a new class
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateClassRequest'
            example:
              name: Grade 5
      responses:
        '201':
          description: Class created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Class'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /admin/classes/{classId}:
    get:
      tags: [Admin — Classes]
      summary: Get a single class with its sections
      parameters:
        - $ref: '#/components/parameters/ClassIdParam'
      responses:
        '200':
          description: Class detail with sections
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Class'
                  - type: object
                    properties:
                      sections:
                        type: array
                        items:
                          $ref: '#/components/schemas/Section'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Admin — Classes]
      summary: Rename a class
      parameters:
        - $ref: '#/components/parameters/ClassIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateClassRequest'
      responses:
        '200':
          description: Class updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Class'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /admin/classes/{classId}/status:
    patch:
      tags: [Admin — Classes]
      summary: Activate or deactivate a class
      parameters:
        - $ref: '#/components/parameters/ClassIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [is_active]
              properties:
                is_active:
                  type: boolean
                  example: false
      responses:
        '200':
          description: Status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Class'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ═══════════════════════════════════════════════════════════════════════════
  # ADMIN — SECTIONS
  # ═══════════════════════════════════════════════════════════════════════════
  /admin/sections:
    get:
      tags: [Admin — Sections]
      summary: List all sections, optionally filtered by class
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: class_id
          in: query
          schema:
            type: string
            format: uuid
        - name: include_inactive
          in: query
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Paginated section list
          content:
            application/json:
              schema:
                type: object
                required: [data, pagination]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Section'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [Admin — Sections]
      summary: Create a new section inside a class
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSectionRequest'
            example:
              class_id: "550e8400-e29b-41d4-a716-446655440000"
              name: Section A
      responses:
        '201':
          description: Section created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Section'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /admin/sections/{sectionId}:
    get:
      tags: [Admin — Sections]
      summary: Get a section with its assigned teachers and students
      parameters:
        - $ref: '#/components/parameters/SectionIdParam'
      responses:
        '200':
          description: Section detail
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Section'
                  - type: object
                    properties:
                      teachers:
                        type: array
                        items:
                          type: object
                          required: [teacher_id, teacher_name, assigned_at]
                          properties:
                            teacher_id:
                              $ref: '#/components/schemas/UUID'
                            teacher_name:
                              type: string
                              example: Mrs. Anitha
                            assigned_at:
                              $ref: '#/components/schemas/Timestamp'
                      students:
                        type: array
                        items:
                          type: object
                          required: [student_id, student_name, assigned_at]
                          properties:
                            student_id:
                              $ref: '#/components/schemas/UUID'
                            student_name:
                              type: string
                              example: Arun Prakash
                            assigned_at:
                              $ref: '#/components/schemas/Timestamp'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Admin — Sections]
      summary: Rename a section
      parameters:
        - $ref: '#/components/parameters/SectionIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSectionRequest'
      responses:
        '200':
          description: Section renamed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Section'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /admin/sections/{sectionId}/status:
    patch:
      tags: [Admin — Sections]
      summary: Activate or deactivate a section
      parameters:
        - $ref: '#/components/parameters/SectionIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [is_active]
              properties:
                is_active:
                  type: boolean
                  example: false
      responses:
        '200':
          description: Status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Section'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ═══════════════════════════════════════════════════════════════════════════
  # ADMIN — USERS
  # ═══════════════════════════════════════════════════════════════════════════
...
  /teacher/sections:
    get:
      tags: [Teacher — Sections]
      summary: Get all sections assigned to the authenticated teacher
      responses:
        '200':
          description: Teacher's assigned sections
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      allOf:
                        - $ref: '#/components/schemas/Section'
                        - type: object
                          properties:
                            lesson_count:
                              type: integer
                              example: 8
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # ═══════════════════════════════════════════════════════════════════════════
  # TEACHER — LESSONS
  # ═══════════════════════════════════════════════════════════════════════════
...